cleanups
git-svn-id: https://svn.code.sf.net/p/linux-gpib/code/trunk@195 b61deed1-901b-0410-a6dd-e94ecbdaea89

+++ linux-gpib/driver/protocol/ibprot.h
@@ -9,4 +9,4 @@ extern int pollTimeidx;
 extern int myPAD;
 extern int mySAD;
 
-extern uint32_t timeTable[];
+extern unsigned long timeTable[];
